+++ src/GafferModule/ValuePlugBinding.cpp
@@ -83,6 +83,12 @@ bool isSetToDefault( ValuePlug *plug )
 	return plug->isSetToDefault();
 }
 
+void resetDefault( ValuePlug *plug )
+{
+	IECorePython::ScopedGILRelease r;
+	plug->resetDefault();
+}
+
 IECore::MurmurHash hash( ValuePlug *plug )
 {
 	// we use a GIL release here to prevent a lock in the case where this triggers a graph
@@ -117,6 +123,7 @@ void GafferModule::bindValuePlug()
 		.def( "setFrom", setFrom )
 		.def( "setToDefault", setToDefault )
 		.def( "isSetToDefault", isSetToDefault )
+		.def( "resetDefault", resetDefault )
 		.def( "hash", hash )
 		.def( "hash", hash2 )
 		.def( "getCacheMemoryLimit", &ValuePlug::getCacheMemoryLimit )