+++ Changes.md
@@ -60,6 +60,7 @@ API
 - ValuePlug
   - Improved interactive performance by not clearing the entire hash cache every time a plug is dirtied. Beware : this can reveal subtle bugs in `DependencyNode::affects()` implementations, causing hashes to be reused if a plug has not been dirtied appropriately. These bugs may previously have gone unnoticed but will now need fixing as a matter of urgency. The GAFFER_CLEAR_HASHCACHE_ON_DIRTY environment variable may be used to enable legacy behaviour in the interim.
   - Added `clearHashCache()` static method.
+  - Added `resetDefault()` method.
 - ScenePlug :
   - Added `existsPlug()` accessor, and deprecated the argumentless overload of the `exists()` method.
   - Added `childBoundsPlug()`, `childBounds()` and `childBoundsHash()` methods.
@@ -146,6 +147,7 @@ Breaking Changes
 - FormatPlug :
   - Removed private member.
   - Changed signature of `defaultValue()` method.
+- ValuePlug : Added virtual method.
 
 Build
 -----